Use Write instead of Create fsnotify event
I noticed that occasionally the cert file would get created and
autosignr would pick up on it before the cert would get written out.
This would lead to a `No PEM data found in block` error, when in fact
the cert just hadn't been written to the file yet. This change now waits
until the file is written to and should now contain the data.